refactor(frontend): features/*/hooks/ の camelCase を use-kebab-case へ揃える（#1040）
本リポの規約 docs/development/frontend-standards.md:668
（hooks: `use-kebab-case.ts`）に対する既存違反 25 件を解消する。

- 25 file を rename。内訳 = **R100 が 21件**（中身バイト不変）／
  **R096〜098 が 4件**（併置 test が自分の実装を `./useXxx` で import しており、
  その1行だけ追随したため。ロジック変更ではない）
- **エクスポート名は不変**（`export function useCommentSection` 等はそのまま）
- 31 file の import を追随（拡張子なし相対参照のパス末尾トークンのみ置換）
- 手作業でなくスクリプトで実施（hub 裁定 2026-08-04 の条件）:
  _work/reports/a1-records/rename-hooks-to-kebab.py

これは A1'（hooks segment → model/ 移設・fleet reg 05:916）のブロッカーだった。
codemod nene2-a1-hooks-to-model は /^use-.+\.tsx?$/ にしかマッチしないため、
リネーム前は 60 件中 34 件しか move されず半分移行になっていた。

受入条件（hub 設定）: リネーム後の --dry-run が features 配下 59/59 move・取り残し 0
→ 実測 59/59・取り残し 0 で PASS。

検証: npm run check --prefix frontend = exit 0
（type-check / lint --max-warnings 0 / prettier / vitest 119 files 666 tests /
validate:themes / knip / build-storybook）。
