+++ frontend/src/features/comment-section/index.ts
@@ -1,3 +1,3 @@
 export { CommentSection } from './ui/CommentSection'
-export { useCommentSection } from './hooks/useCommentSection'
-export type { CommentSectionState } from './hooks/useCommentSection'
+export { useCommentSection } from './hooks/use-comment-section'
+export type { CommentSectionState } from './hooks/use-comment-section'
