+++ frontend/src/features/manage-appearance/hooks/use-floating-cta-page.ts
