+++ frontend/src/features/manage-appearance/hooks/use-footer-config-page.ts
