+++ frontend/src/features/manage-appearance/hooks/use-home-hero-page.ts
