+++ frontend/src/features/manage-appearance/hooks/use-permalink-settings-page.ts
