+++ frontend/src/features/manage-appearance/hooks/use-theme-preview-sender.ts
