+++ frontend/src/features/manage-appearance/index.ts
@@ -1,17 +1,17 @@
 export { AppearanceView } from './ui/AppearanceView'
 export { PermalinkSettingsView } from './ui/PermalinkSettingsView'
-export { usePermalinkSettingsPage } from './hooks/usePermalinkSettingsPage'
-export type { PermalinkSettingsPageState } from './hooks/usePermalinkSettingsPage'
+export { usePermalinkSettingsPage } from './hooks/use-permalink-settings-page'
+export type { PermalinkSettingsPageState } from './hooks/use-permalink-settings-page'
 export { ThemeWorkspace } from './ui/workspace/ThemeWorkspace'
-export { usePublicThemePage } from './hooks/usePublicThemePage'
-export type { PublicThemePageState } from './hooks/usePublicThemePage'
-export { useThemeCustomizePage } from './hooks/useThemeCustomizePage'
-export type { ThemeCustomizePageState } from './hooks/useThemeCustomizePage'
-export { useFooterConfigPage } from './hooks/useFooterConfigPage'
-export { useHeaderConfigPage } from './hooks/useHeaderConfigPage'
-export type { HeaderConfigPageState } from './hooks/useHeaderConfigPage'
-export { useHomeHeroPage } from './hooks/useHomeHeroPage'
-export type { HomeHeroPageState } from './hooks/useHomeHeroPage'
+export { usePublicThemePage } from './hooks/use-public-theme-page'
+export type { PublicThemePageState } from './hooks/use-public-theme-page'
+export { useThemeCustomizePage } from './hooks/use-theme-customize-page'
+export type { ThemeCustomizePageState } from './hooks/use-theme-customize-page'
+export { useFooterConfigPage } from './hooks/use-footer-config-page'
+export { useHeaderConfigPage } from './hooks/use-header-config-page'
+export type { HeaderConfigPageState } from './hooks/use-header-config-page'
+export { useHomeHeroPage } from './hooks/use-home-hero-page'
+export type { HomeHeroPageState } from './hooks/use-home-hero-page'
 export { FloatingCtaView } from './ui/FloatingCtaView'
-export { useFloatingCtaPage } from './hooks/useFloatingCtaPage'
-export type { FloatingCtaPageState } from './hooks/useFloatingCtaPage'
+export { useFloatingCtaPage } from './hooks/use-floating-cta-page'
+export type { FloatingCtaPageState } from './hooks/use-floating-cta-page'
