+++ frontend/src/features/manage-appearance/ui/FloatingCtaView.tsx
@@ -10,7 +10,7 @@ import {
   safeHref,
 } from '@/shared/lib/floating-cta'
 import { Button, Card, Stack, Text } from '@/shared/ui'
-import type { FloatingCtaPageState } from '../hooks/useFloatingCtaPage'
+import type { FloatingCtaPageState } from '../hooks/use-floating-cta-page'
 
 function Row({ label, children }: { label: string; children: ReactNode }) {
   return (
