+++ frontend/src/features/manage-appearance/ui/HomeHeroView.tsx
@@ -1,7 +1,7 @@
 import { BlocksFieldEditor } from '@/features/edit-entity-text-fields'
 import { useTranslation } from '@/shared/i18n'
 import { Button, LoadingState, Stack, Text } from '@/shared/ui'
-import type { HomeHeroPageState } from '../hooks/useHomeHeroPage'
+import type { HomeHeroPageState } from '../hooks/use-home-hero-page'
 
 const HERO_ONLY = ['hero'] as const
