+++ frontend/src/features/manage-appearance/ui/workspace/ThemeWorkspace.test.tsx
@@ -7,12 +7,12 @@ import { DEFAULT_HEADER_CONFIG } from '@/shared/lib/header-config'
 import { PUBLIC_THEMES } from '@/shared/lib/public-themes'
 import { mswServer } from '@tests/msw/server'
 import { renderWithProviders } from '@tests/render/render-with-providers'
-import type { FloatingCtaPageState } from '../../hooks/useFloatingCtaPage'
-import type { FooterConfigPageState } from '../../hooks/useFooterConfigPage'
-import type { HeaderConfigPageState } from '../../hooks/useHeaderConfigPage'
-import type { HomeHeroPageState } from '../../hooks/useHomeHeroPage'
-import type { PublicThemePageState } from '../../hooks/usePublicThemePage'
-import type { ThemeCustomizePageState } from '../../hooks/useThemeCustomizePage'
+import type { FloatingCtaPageState } from '../../hooks/use-floating-cta-page'
+import type { FooterConfigPageState } from '../../hooks/use-footer-config-page'
+import type { HeaderConfigPageState } from '../../hooks/use-header-config-page'
+import type { HomeHeroPageState } from '../../hooks/use-home-hero-page'
+import type { PublicThemePageState } from '../../hooks/use-public-theme-page'
+import type { ThemeCustomizePageState } from '../../hooks/use-theme-customize-page'
 import { ThemeWorkspace } from './ThemeWorkspace'
 
 function makeStates(
