+++ frontend/src/features/manage-appearance/ui/workspace/ThemeWorkspace.tsx
@@ -5,13 +5,13 @@ import { useTranslation, type MessageKey } from '@/shared/i18n'
 import { resolveDraftImageUrls, type ThemeImages } from '@/shared/lib/theme-customization'
 import { THEME_PREVIEW_PARAM } from '@/shared/lib/theme-preview-protocol'
 import { Button, Text } from '@/shared/ui'
-import type { FloatingCtaPageState } from '../../hooks/useFloatingCtaPage'
-import type { FooterConfigPageState } from '../../hooks/useFooterConfigPage'
-import type { HeaderConfigPageState } from '../../hooks/useHeaderConfigPage'
-import type { HomeHeroPageState } from '../../hooks/useHomeHeroPage'
-import type { PublicThemePageState } from '../../hooks/usePublicThemePage'
-import type { ThemeCustomizePageState } from '../../hooks/useThemeCustomizePage'
-import { useThemePreviewSender } from '../../hooks/useThemePreviewSender'
+import type { FloatingCtaPageState } from '../../hooks/use-floating-cta-page'
+import type { FooterConfigPageState } from '../../hooks/use-footer-config-page'
+import type { HeaderConfigPageState } from '../../hooks/use-header-config-page'
+import type { HomeHeroPageState } from '../../hooks/use-home-hero-page'
+import type { PublicThemePageState } from '../../hooks/use-public-theme-page'
+import type { ThemeCustomizePageState } from '../../hooks/use-theme-customize-page'
+import { useThemePreviewSender } from '../../hooks/use-theme-preview-sender'
 import { FloatingCtaView } from '../FloatingCtaView'
 import { FooterContentView } from '../FooterContentView'
 import { HeaderContentView } from '../HeaderContentView'
