+++ frontend/src/features/manage-entity-status/hooks/use-entity-seo-panel.ts
