+++ frontend/src/features/manage-entity-status/hooks/use-entity-status-panel.ts
