+++ frontend/src/features/manage-entity-status/index.ts
@@ -1,9 +1,9 @@
 export { EntityRevisionsPanel } from './ui/EntityRevisionsPanel'
 export { EntitySeoPanel } from './ui/EntitySeoPanel'
 export { EntityStatusPanel } from './ui/EntityStatusPanel'
-export { useEntitySeoPanel } from './hooks/useEntitySeoPanel'
-export { useEntityRevisionsPanel } from './hooks/useEntityRevisionsPanel'
-export { useEntityStatusPanel } from './hooks/useEntityStatusPanel'
-export type { EntitySeoPanelState } from './hooks/useEntitySeoPanel'
-export type { EntityRevisionsPanelState } from './hooks/useEntityRevisionsPanel'
-export type { EntityStatusPanelState } from './hooks/useEntityStatusPanel'
+export { useEntitySeoPanel } from './hooks/use-entity-seo-panel'
+export { useEntityRevisionsPanel } from './hooks/use-entity-revisions-panel'
+export { useEntityStatusPanel } from './hooks/use-entity-status-panel'
+export type { EntitySeoPanelState } from './hooks/use-entity-seo-panel'
+export type { EntityRevisionsPanelState } from './hooks/use-entity-revisions-panel'
+export type { EntityStatusPanelState } from './hooks/use-entity-status-panel'
