+++ frontend/src/features/manage-front-page/hooks/use-front-page.ts
