+++ frontend/src/features/manage-front-page/index.ts
@@ -1,8 +1,8 @@
-export { useFrontPage } from './hooks/useFrontPage'
+export { useFrontPage } from './hooks/use-front-page'
 export type {
   FrontPageMode,
   FrontPageRecordOption,
   FrontPageState,
   FrontPageTypeOption,
-} from './hooks/useFrontPage'
+} from './hooks/use-front-page'
 export { FrontPageView } from './ui/FrontPageView'
