+++ frontend/src/features/manage-notification-channels/index.ts
@@ -1,3 +1,3 @@
 export { ManageNotificationChannelsView } from './ui/ManageNotificationChannelsView'
-export { useManageNotificationChannelsPage } from './hooks/useManageNotificationChannelsPage'
-export type { ManageNotificationChannelsPageState } from './hooks/useManageNotificationChannelsPage'
+export { useManageNotificationChannelsPage } from './hooks/use-manage-notification-channels-page'
+export type { ManageNotificationChannelsPageState } from './hooks/use-manage-notification-channels-page'
