+++ frontend/src/features/manage-organizations/hooks/use-manage-organization-detail-page.ts
