+++ frontend/src/features/manage-organizations/index.ts
@@ -1,6 +1,6 @@
 export { ManageOrganizationsView } from './ui/ManageOrganizationsView'
 export { ManageOrganizationDetailView } from './ui/ManageOrganizationDetailView'
-export { useManageOrganizationsPage } from './hooks/useManageOrganizationsPage'
-export { useManageOrganizationDetailPage } from './hooks/useManageOrganizationDetailPage'
-export type { ManageOrganizationsPageState } from './hooks/useManageOrganizationsPage'
-export type { ManageOrganizationDetailPageState } from './hooks/useManageOrganizationDetailPage'
+export { useManageOrganizationsPage } from './hooks/use-manage-organizations-page'
+export { useManageOrganizationDetailPage } from './hooks/use-manage-organization-detail-page'
+export type { ManageOrganizationsPageState } from './hooks/use-manage-organizations-page'
+export type { ManageOrganizationDetailPageState } from './hooks/use-manage-organization-detail-page'
