+++ frontend/src/features/manage-record-page/hooks/use-record-page-display.ts
