+++ frontend/src/features/manage-record-page/index.ts
@@ -1,3 +1,3 @@
-export { useRecordPageDisplay } from './hooks/useRecordPageDisplay'
-export type { RecordPageDisplayState } from './hooks/useRecordPageDisplay'
+export { useRecordPageDisplay } from './hooks/use-record-page-display'
+export type { RecordPageDisplayState } from './hooks/use-record-page-display'
 export { RecordPageDisplayView } from './ui/RecordPageDisplayView'
