+++ frontend/src/features/manage-settings/index.ts
@@ -1,3 +1,3 @@
 export { ManageSiteSettingsView } from './ui/ManageSiteSettingsView'
-export { useManageSiteSettingsPage } from './hooks/useManageSiteSettingsPage'
-export type { ManageSiteSettingsPageState } from './hooks/useManageSiteSettingsPage'
+export { useManageSiteSettingsPage } from './hooks/use-manage-site-settings-page'
+export type { ManageSiteSettingsPageState } from './hooks/use-manage-site-settings-page'
