+++ frontend/src/features/manage-webhooks/index.ts
@@ -1,3 +1,3 @@
 export { ManageWebhooksView } from './ui/ManageWebhooksView'
-export { useManageWebhooksPage } from './hooks/useManageWebhooksPage'
-export type { ManageWebhooksPageState } from './hooks/useManageWebhooksPage'
+export { useManageWebhooksPage } from './hooks/use-manage-webhooks-page'
+export type { ManageWebhooksPageState } from './hooks/use-manage-webhooks-page'
